requirements.txt
beautifulsoup4==4.6.0
certifi==2017.11.5
chardet==3.0.4
click==6.7
Flask==0.12.2
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
pdfminer.six==20170720
peewee==2.10.2
psycopg2==2.7.3.2
pycryptodome==3.4.7
PyYAML==3.12
requests==2.18.4
six==1.11.0
slackclient==1.0.9
urllib3==1.22
virtualenv==15.1.0
websocket-client==0.44.0
Werkzeug==0.12.2